Update sqlite_driver.py and rename to sql_driver.py
